Keep list of already dumped structs to prevent duplications
I was getting multiple outputs of list_head when running this on <linux/list.h>.
Not sure why. Or how I can filter only the real struct declaration, as there must be
only a single one...
This works fine, however, and it will be helpful when I'll recursively dump structs.