Handle user-types typedefs
This fixes handling of types like "atomic_t" in Linux, which
have no "ORIG_TYPE_NAME" because they are defined this way:
"typedef struct { ... } atomic_t".

Whenever a type has an underlying type name (that is, a named
TYPE_MAIN_VARIANT) then we'll use that. Otherwise, resort to using the
TYPE_IDENTIFIER which might be a typedef.